Reset connection when there's a peer reset. Also refactored the code structure
