+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = bosch-control-panel-cc880p
-version = 1.0.2
+version = 1.1.0
 author = Hugo Gomes
 author_email = hgomes88@gmail.com
 description = Library to interface with the old CC880p Bosch COntrol Panels
@@ -17,7 +17,7 @@ classifiers =
 [options]
 package_dir =
     = src
-packages = find:
+packages = find_namespace:
 python_requires = >=3.6
 install_requires = 
     aioretry==5.0.2
