+++ src/bosch/control_panel/cc880p/models.py
