+++ src/bosch/utils/bytes_to_str.py
